app/requirements.txt
elasticsearch
Flask==1.1.4
Flask-SQLAlchemy
pytest
sklearn
gensim
Jinja2
SQLAlchemy
uwsgi
mecab-python3==0.996.5